+++ intrastat_base/country_view.xml
@@ -9,8 +9,8 @@
 <data>
 
 <!-- Add intrastat field on res_country -->
-<record id="fr_intrastat_base_country_tree" model="ir.ui.view">
-    <field name="name">fr.intrastat.base.country.tree</field>
+<record id="intrastat_base_country_tree" model="ir.ui.view">
+    <field name="name">intrastat.base.country.tree</field>
     <field name="model">res.country</field>
     <field name="inherit_id" ref="base.view_country_tree" />
     <field name="arch" type="xml">
@@ -20,8 +20,8 @@
     </field>
 </record>
 
-<record id="fr_intrastat_base_country_form" model="ir.ui.view">
-    <field name="name">fr.intrastat.base.country.form</field>
+<record id="intrastat_base_country_form" model="ir.ui.view">
+    <field name="name">intrastat.base.country.form</field>
     <field name="model">res.country</field>
     <field name="inherit_id" ref="base.view_country_form" />
     <field name="arch" type="xml">
