+++ intrastat_base/product_demo.xml
@@ -1,0 +1,22 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<!--
+    Copyright (C) 2011 Akretion (http://www.akretion.com/)
+    @author Alexis de Lattre <alexis.delattre@akretion.com>
+    The licence is in the file __terp__.py
+-->
+
+<openerp>
+<data noupdate="1">
+
+<record id="shipping_costs_exclude" model="product.product">
+    <field name="name">Shipping costs</field>
+    <field name="code">SHIP</field>
+    <field name="type">service</field>
+    <field name="categ_id" ref="product.product_category_services"/>
+    <field name="list_price">30</field>
+    <field name="exclude_from_intrastat">False</field>
+</record>
+
+</data>
+</openerp>
