--- l10n_fr_intrastat_base/account_invoice_workflow.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<openerp>
-
-	<data>
-		<record id="act_legal_intrastat" model="workflow.activity">
-			<field name="wkf_id" ref="account.wkf" />
-			<field name="name">legal_intrastat</field>
-			<field name="action">write({'state':'legal_intrastat'})</field>
-			<field name="kind">function</field>
-		</record>
-
-		<record id="t1" model="workflow.transition">
-			<field name="act_from" ref="account.act_draft" />
-			<field name="act_to" ref="act_legal_intrastat" />
-			<field name="signal">invoice_legal_intrastat</field>
-		</record>
-
-		<record id="t2" model="workflow.transition">
-			<field name="act_from" ref="act_legal_intrastat" />
-			<field name="act_to" ref="account.act_cancel" />
-			<field name="signal">invoice_cancel</field>
-		</record>
-	</data>
-</openerp>
