--- l10n_fr_intrastat_base/company_view.xml
@@ -1,27 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<!--
-    Copyright (C) 2010-2011 Akretion (http://www.akretion.com/)
-    @author Alexis de Lattre <alexis.delattre@akretion.com>
-    The licence is in the file __terp__.py
--->
-
-<openerp>
-<data>
-
-<record id="fr_intrastat_base_company_form" model="ir.ui.view">
-    <field name="name">fr.intrastat.base.company.form</field>
-    <field name="model">res.company</field>
-    <field name="inherit_id" ref="base.view_company_form" />
-    <field name="arch" type="xml">
-        <notebook>
-            <page position="inside" string="France intrastat settings">
-                <separator colspan="4" string="Intrastat Type" />
-                <field nolabel="1" colspan="4" name="intrastat_type_ids" />
-            </page>
-        </notebook>
-    </field>
-</record>
-
-</data>
-</openerp>
