--- l10n_fr_intrastat_base/intrastat_type_data.xml
@@ -1,31 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<!--
-    Copyright (C) 2010-2011 Akretion (http://www.akretion.com/)
-    @author Alexis de Lattre <alexis.delattre@akretion.com>
-    The licence is in the file __terp__.py
--->
-
-<openerp>
-<data noupdate="1">
-
-<record id="regular_export" model="report.intrastat.type">
-    <field name="name">Regular export</field>
-    <field name="company_id" ref="base.main_company"/>
-    <field name="type">export</field>
-</record>
-
-<record id="regular_import" model="report.intrastat.type">
-    <field name="name">Regular import</field>
-    <field name="company_id" ref="base.main_company"/>
-    <field name="type">import</field>
-</record>
-
-<record id="other" model="report.intrastat.type">
-     <field name="name">Exclude from intrastat (special)</field>
-     <field name="company_id" ref="base.main_company"/>
-     <field name="type">other</field>
-</record>
-
-</data>
-</openerp>
